file reorg, basic profile view ui setup, settings button in profile toolbar
