+++ Pachanga/Core/Authentication/AuthenticationView.swift
@@ -9,26 +9,6 @@ import GoogleSignIn
 import GoogleSignInSwift
 import SwiftUI
 
-
-
-@MainActor
-final class AuthenticationViewModel: ObservableObject {
-    
-    let signInAppleHelper = SignInAppleHelper()
-    
-    func signInGoogle() async throws {
-        let helper = SignInGoogleHelper()
-        let tokens = try await helper.signIn()
-        try await AuthenticationManager.shared.signInWithGoogle(tokens: tokens)
-    }
-    
-    func signInApple() async throws {
-        let helper = SignInAppleHelper()
-        let tokens = try await helper.startSignInWithAppleFlow()
-        try await AuthenticationManager.shared.signInWithApple(tokens: tokens)
-    }
-}
-
 struct AuthenticationView: View {
     
     @StateObject private var viewModel = AuthenticationViewModel()
