+++ Pachanga/Core/Authentication/AuthenticationViewModel.swift
@@ -1,0 +1,26 @@
+//
+//  AuthenticationViewModel.swift
+//  Pachanga
+//
+//  Created by Javier Alaves on 12/8/23.
+//
+
+import Foundation
+
+@MainActor
+final class AuthenticationViewModel: ObservableObject {
+    
+    let signInAppleHelper = SignInAppleHelper()
+    
+    func signInGoogle() async throws {
+        let helper = SignInGoogleHelper()
+        let tokens = try await helper.signIn()
+        try await AuthenticationManager.shared.signInWithGoogle(tokens: tokens)
+    }
+    
+    func signInApple() async throws {
+        let helper = SignInAppleHelper()
+        let tokens = try await helper.startSignInWithAppleFlow()
+        try await AuthenticationManager.shared.signInWithApple(tokens: tokens)
+    }
+}
