+++ Pachanga/Core/Authentication/Subviews/SignInEmailView.swift
@@ -7,36 +7,6 @@
 
 import SwiftUI
 
-@MainActor
-final class SignInEmailViewModel: ObservableObject {
-    
-    @Published var email = ""
-    @Published var password = ""
-    
-    func signUp() async throws {
-        
-        // Do proper validation down the line
-        guard !email.isEmpty, !password.isEmpty else {
-            print("No email or password found")
-            return
-        }
-        
-        try await AuthenticationManager.shared.createUser(email: email, password: password)
-    }
-    
-    func signIn() async throws {
-        
-        // Do proper validation down the line
-        guard !email.isEmpty, !password.isEmpty else {
-            print("No email or password found")
-            return
-        }
-        
-        try await AuthenticationManager.shared.signInUser(email: email, password: password)
-    }
-    
-}
-
 struct SignInEmailView: View {
     
     @StateObject private var viewModel = SignInEmailViewModel()
