+++ Pachanga/Core/Authentication/Subviews/SignInEmailViewModel.swift
@@ -1,0 +1,38 @@
+//
+//  SignInEmailViewModel.swift
+//  Pachanga
+//
+//  Created by Javier Alaves on 12/8/23.
+//
+
+import Foundation
+
+@MainActor
+final class SignInEmailViewModel: ObservableObject {
+    
+    @Published var email = ""
+    @Published var password = ""
+    
+    func signUp() async throws {
+        
+        // Do proper validation down the line
+        guard !email.isEmpty, !password.isEmpty else {
+            print("No email or password found")
+            return
+        }
+        
+        try await AuthenticationManager.shared.createUser(email: email, password: password)
+    }
+    
+    func signIn() async throws {
+        
+        // Do proper validation down the line
+        guard !email.isEmpty, !password.isEmpty else {
+            print("No email or password found")
+            return
+        }
+        
+        try await AuthenticationManager.shared.signInUser(email: email, password: password)
+    }
+    
+}
