+++ Pachanga/Core/Profile/ProfileView.swift
@@ -1,0 +1,56 @@
+//
+//  ProfileView.swift
+//  Pachanga
+//
+//  Created by Javier Alaves on 7/7/23.
+//
+
+import SwiftUI
+
+@MainActor
+final class ProfileViewModel: ObservableObject {
+    
+    @Published private(set) var user: AuthDataResultModel? = nil
+    
+    func loadCurrentUser() throws {
+        self.user = try AuthenticationManager.shared.getAuthenticatedUser()
+    }
+    
+}
+
+struct ProfileView: View {
+    
+    @StateObject private var viewModel = ProfileViewModel()
+    @Binding var showSignInView: Bool
+    
+    var body: some View {
+        List {
+            if let user = viewModel.user {
+                Text("ID: \(user.uid)")
+            }
+        }
+        .onAppear {
+            try? viewModel.loadCurrentUser()
+        }
+        .navigationTitle("Perfil")
+        .toolbar {
+            ToolbarItem(placement: .navigationBarTrailing) {
+                
+                NavigationLink {
+                    SettingsView(showSignInView: $showSignInView)
+                } label: {
+                    Image(systemName: "gear")
+                        .font(.headline)
+                }
+            }
+        }
+    }
+}
+
+struct ProfileView_Previews: PreviewProvider {
+    static var previews: some View {
+        NavigationStack {
+            ProfileView(showSignInView: .constant(false))
+        }
+    }
+}
