+++ Pachanga/Core/Settings/SettingsView.swift
@@ -7,23 +7,6 @@
 
 import SwiftUI
 
-@MainActor
-final class SettingsViewModel: ObservableObject {
-    
-    @Published var authProviders: [AuthProviderOption] = []
-    
-    func loadAuthProviders() {
-        if let providers = try? AuthenticationManager.shared.getProviders() {
-            authProviders = providers
-        }
-    }
-    
-    func signOut() throws {
-        try AuthenticationManager.shared.signOut()
-    }
-    
-}
-
 struct SettingsView: View {
     
     @StateObject private var viewModel = SettingsViewModel()
