+++ Pachanga/Core/Settings/SettingsViewModel.swift
@@ -1,0 +1,25 @@
+//
+//  SettingsViewModel.swift
+//  Pachanga
+//
+//  Created by Javier Alaves on 12/8/23.
+//
+
+import Foundation
+
+@MainActor
+final class SettingsViewModel: ObservableObject {
+    
+    @Published var authProviders: [AuthProviderOption] = []
+    
+    func loadAuthProviders() {
+        if let providers = try? AuthenticationManager.shared.getProviders() {
+            authProviders = providers
+        }
+    }
+    
+    func signOut() throws {
+        try AuthenticationManager.shared.signOut()
+    }
+    
+}
