+++ Pachanga/PachangaApp.swift
@@ -17,7 +17,7 @@ struct PachangaApp: App {
     
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            MainView()
         }
     }
 }
