--- Pachanga/ProfileView.swift
@@ -1,22 +1,0 @@
-//
-//  ProfileView.swift
-//  Pachanga
-//
-//  Created by Javier Alaves on 7/7/23.
-//
-
-import SwiftUI
-
-struct ProfileView: View {
-    var body: some View {
-        Text("Profile")
-            .font(.largeTitle)
-            .fontWeight(.semibold)
-    }
-}
-
-struct ProfileView_Previews: PreviewProvider {
-    static var previews: some View {
-        ProfileView()
-    }
-}
